Refactor: Encapsulate file handling
